pyproject.toml
[tool.poetry]
name = "wv-render"
version = "0.1.0"
description = ""
authors = ["cynthia2006 <121848440+cynthia2006@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python        = "^3.12"
pyFFTW        = "^0.14.0"
av            = "^12.3.0"
skia-python   = "^87.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
